用户微服务子模块，添加token验证

+++ tensquare_parent/tensquare_user/src/main/java/com/tensquare/user/controller/UserController.java
@@ -1,7 +1,9 @@
 package com.tensquare.user.controller;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import io.jsonwebtoken.Claims;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -17,6 +19,10 @@ import com.tensquare.user.service.UserService;
 import entity.PageResult;
 import entity.Result;
 import entity.StatusCode;
+import util.JwtUtil;
+
+import javax.servlet.http.HttpServletRequest;
+
 /**
  * user控制器层
  * @author Administrator
@@ -29,6 +35,11 @@ public class UserController {
 
 	@Autowired
 	private UserService userService;
+
+	@Autowired
+	private HttpServletRequest request;
+	@Autowired
+	private JwtUtil jwtUtil;
 	
 	
 	/**
@@ -94,13 +105,17 @@ public class UserController {
 		userService.update(user);
 		return new Result(true,StatusCode.OK,"修改成功",null);
 	}
-	
+
 	/**
 	 * 删除
 	 * @param id
 	 */
 	@RequestMapping(value="/{id}",method= RequestMethod.DELETE)
-	public Result delete(@PathVariable String id){
+	public Result delete(@PathVariable String id ){
+		Claims claims=(Claims) request.getAttribute("admin_claims");
+		if(claims==null){
+			return new Result(true,StatusCode.ACCESSERROR,"无权访问",null);
+		}
 		userService.deleteById(id);
 		return new Result(true,StatusCode.OK,"删除成功",null);
 	}
@@ -123,21 +138,28 @@ public class UserController {
 		userService.add(user,code);
 		return new Result(true,StatusCode.OK,"注册成功",null);
 	}
+
 	/**
-	 * 用户登陆
-	 * @param mobile
-	 * @param password
+	 * 用户登录
+	 * @param loginMap
 	 * @return
 	 */
 	@RequestMapping(value="/login",method=RequestMethod.POST)
-	public Result login(String mobile,String password){
-		User user = userService.findByMobileAndPassword(mobile,password);
+	public Result login(@RequestBody Map<String,String> loginMap){
+		User user =
+				userService.findByMobileAndPassword(loginMap.get("mobile"),loginMap.get("password"));
 		if(user!=null){
-			return new Result(true,StatusCode.OK,"登陆成功",null);
+			String token = jwtUtil.createJWT(user.getId(), user.getNickname(), "user");//签发token
+			Map map=new HashMap();
+			map.put("token",token);
+			map.put("name",user.getNickname());//昵称
+			map.put("avatar",user.getAvatar());//头像
+			return new Result(true,StatusCode.OK,"登陆成功",map);
 		}else{
 			return new Result(false,StatusCode.LOGINERROE,"用户名或密码错误",null);
 		}
 	}
+
